DR-409 Reconfiguring the structure
