+++ scripts/integration-tests.sh
@@ -4,18 +4,14 @@
 set -e
 # This script runs python integration tests
 #
-APPLICATION_DIR=application
 APPLICATION_TEST_DIR='tests/e2e-api'
 
 source ./scripts/functions/git-functions.sh
 
+# identify the workspace name and set TERRAFORM_WORKSPACE_NAME
 export_terraform_workspace_name
 
-
-pip install -r $APPLICATION_TEST_DIR/requirements.txt
-# check exports have been done
-
-# check exports have been done
+# check export has been done
 EXPORTS_SET=0
 if [ -z "$TERRAFORM_WORKSPACE_NAME" ] ; then
   echo Set TERRAFORM_WORKSPACE_NAME
@@ -23,9 +19,13 @@ if [ -z "$TERRAFORM_WORKSPACE_NAME" ] ; then
 fi
 
 if [ $EXPORTS_SET = 1 ] ; then
-  echo One or more exports not set
+  echo TERRAFORM_WORKSPACE_NAME exports not set
   exit 1
 fi
+
+# install requirements
+echo "Installing requirements"
+pip install -r $APPLICATION_TEST_DIR/requirements.txt
 
 echo "Running tests for E2E api"
 # echo $TERRAFORM_WORKSPACE_NAME
