+++ tests/e2e-api/behave.ini
@@ -5,5 +5,5 @@ stderr_capture = no
 log_capture = no
 show_skipped = false
 [behave.userdata]
-workspace=
-apigateway=cm-rest-apig
+workspace =
+apigateway = cm-rest-apig
