--- tests/e2e-api/features/environment.py
@@ -1,7 +1,0 @@
-# -- BEHAVE HOOKS:
-def before_all(context):
-    print("Before all executed")
-    userdata = context.config.userdata
-    context.workspace = userdata.get("workspace")
-    context.apigateway = userdata.get("apigateway")
-    print(context.workspace)
